requirements.txt
# Core dependencies
SQLAlchemy==2.0.27
click==8.1.7
python-dotenv==1.0.1
sentry-sdk==1.39.1
Werkzeug==3.0.1  # Pour le hachage des mots de passe

# Development dependencies
black==24.1.1  # Formatage du code
flake8==7.0.0  # Linting
Faker==23.2.1  # Génération de données de test 